+++ src/assetbundles/hommformcp/HOMMFormCPAsset.php
@@ -5,7 +5,7 @@
  * Show form requests in the control panel
  *
  * @link      https://github.com/HOMMinteractive
- * @copyright Copyright (c) 2019 HOMM interactive
+ * @copyright Copyright (c) 2026 HOMM interactive
  */
 
 namespace homm\hommform\assetbundles\hommformcp;
@@ -14,9 +14,9 @@ use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 
 /**
- * @author    Domenik Hofer
+ * @author    Benjamin Ammann
  * @package   HOMMForm
- * @since     1.0.0
+ * @since     4.0.0
  */
 class HOMMFormCPAsset extends AssetBundle
 {