+++ src/controllers/ViewerController.php
@@ -5,7 +5,7 @@
  * Show form requests in the control panel
  *
  * @link      https://github.com/HOMMinteractive
- * @copyright Copyright (c) 2019 HOMM interactive
+ * @copyright Copyright (c) 2026 HOMM interactive
  */
 
 namespace homm\hommform\controllers;
@@ -15,9 +15,9 @@ use craft\web\Controller;
 use homm\hommform\HOMMForm;
 
 /**
- * @author    Domenik Hofer
+ * @author    Benjamin Ammann
  * @package   HOMMForm
- * @since     0.0.1
+ * @since     4.0.0
  */
 class ViewerController extends Controller
 {