+++ src/services/ViewerService.php
@@ -5,7 +5,7 @@
  * Show form requests in the control panel
  *
  * @link      https://github.com/HOMMinteractive
- * @copyright Copyright (c) 2019 HOMM interactive
+ * @copyright Copyright (c) 2026 HOMM interactive
  */
 
 namespace homm\hommform\services;
@@ -14,9 +14,9 @@ use craft\db\Query;
 use craft\base\Component;
 
 /**
- * @author    Domenik Hofer
+ * @author    Benjamin Ammann
  * @package   HOMMForm
- * @since     1.0.0
+ * @since     4.0.0
  */
 class ViewerService extends Component
 {