+++ src/variables/HOMMFormVariable.php
@@ -5,7 +5,7 @@
  * Show form requests in the control panel
  *
  * @link      https://github.com/HOMMinteractive
- * @copyright Copyright (c) 2019 HOMM interactive
+ * @copyright Copyright (c) 2026 HOMM interactive
  */
 
 namespace homm\hommform\variables;
@@ -14,9 +14,9 @@ namespace homm\hommform\variables;
 use homm\hommform\HOMMForm;
 
 /**
- * @author    Domenik Hofer
- * @package   Hommjuicer
- * @since     0.0.1
+ * @author    Benjamin Ammann
+ * @package   HOMMForm
+ * @since     4.0.0
  */
 class HOMMFormVariable
 {