+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres
 to [Semantic Versioning](http://semver.org/).
 
+## 4.0.0 - 2026-02-27
+
+- renamed plugin
+- added form submit feature
+- added control panel settings
+
 ## 3.0.2 - 2025-10-24
 
 - added a display option for json's in table